FIX Accountancy - Register accountancy_code for thirdparty & disable completely "product account on thirdparty"

+++ htdocs/societe/class/societe.class.php
@@ -886,9 +886,11 @@ class Societe extends CommonObject
 			$sql .= ", import_key";
 			$sql .= ", fk_multicurrency";
 			$sql .= ", multicurrency_code";
-			if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-				$sql .= ", accountancy_code_buy";
-				$sql .= ", accountancy_code_sell";
+			if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+				if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+					$sql .= ", accountancy_code_buy";
+					$sql .= ", accountancy_code_sell";
+				}
 			}
 			$sql .= ") VALUES ('".$this->db->escape($this->name)."', '".$this->db->escape($this->name_alias)."', ".$this->db->escape($this->entity).", '".$this->db->idate($now)."'";
 			$sql .= ", ".(!empty($user->id) ? ((int) $user->id) : "null");
@@ -902,9 +904,11 @@ class Societe extends CommonObject
 			$sql .= ", ".(!empty($this->import_key) ? "'".$this->db->escape($this->import_key)."'" : "null");
 			$sql .= ", ".(int) $this->fk_multicurrency;
 			$sql .= ", '".$this->db->escape($this->multicurrency_code)."'";
-			if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-				$sql .= ", '" . $this->db->escape($this->accountancy_code_buy) . "'";
-				$sql .= ", '" . $this->db->escape($this->accountancy_code_sell) . "'";
+			if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+				if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+					$sql .= ", '" . $this->db->escape($this->accountancy_code_buy) . "'";
+					$sql .= ", '" . $this->db->escape($this->accountancy_code_sell) . "'";
+				}
 			}
 			$sql .= ")";
 
@@ -916,24 +920,26 @@ class Societe extends CommonObject
 				$ret = $this->update($this->id, $user, 0, 1, 1, 'add');
 
 				// update accountancy for this entity
-				if (!$error && !empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-					$this->db->query("DELETE FROM " . MAIN_DB_PREFIX . "societe_perentity WHERE fk_soc = " . $this->id . " AND entity = " . $conf->entity);
+				if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+					if (!$error && !empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+						$this->db->query("DELETE FROM " . MAIN_DB_PREFIX . "societe_perentity WHERE fk_soc = " . $this->id . " AND entity = " . $conf->entity);
 
-					$sql = "INSERT INTO " . MAIN_DB_PREFIX . "societe_perentity (";
-					$sql .= " fk_soc";
-					$sql .= ", entity";
-					$sql .= ", accountancy_code_buy";
-					$sql .= ", accountancy_code_sell";
-					$sql .= ") VALUES (";
-					$sql .= $this->id;
-					$sql .= ", " . $conf->entity;
-					$sql .= ", '" . $this->db->escape($this->accountancy_code_buy) . "'";
-					$sql .= ", '" . $this->db->escape($this->accountancy_code_sell) . "'";
-					$sql .= ")";
-					$result = $this->db->query($sql);
-					if (!$result) {
-						$error++;
-						$this->error = 'ErrorFailedToUpdateAccountancyForEntity';
+						$sql = "INSERT INTO " . MAIN_DB_PREFIX . "societe_perentity (";
+						$sql .= " fk_soc";
+						$sql .= ", entity";
+						$sql .= ", accountancy_code_buy";
+						$sql .= ", accountancy_code_sell";
+						$sql .= ") VALUES (";
+						$sql .= $this->id;
+						$sql .= ", " . $conf->entity;
+						$sql .= ", '" . $this->db->escape($this->accountancy_code_buy) . "'";
+						$sql .= ", '" . $this->db->escape($this->accountancy_code_sell) . "'";
+						$sql .= ")";
+						$result = $this->db->query($sql);
+						if (!$result) {
+							$error++;
+							$this->error = 'ErrorFailedToUpdateAccountancyForEntity';
+						}
 					}
 				}
 
@@ -1447,9 +1453,11 @@ class Societe extends CommonObject
 			$sql .= ",order_min_amount= ".($this->order_min_amount != '' ? $this->order_min_amount : 'null');
 			$sql .= ",supplier_order_min_amount= ".($this->supplier_order_min_amount != '' ? $this->supplier_order_min_amount : 'null');
 			$sql .= ",fk_prospectlevel='".$this->db->escape($this->fk_prospectlevel)."'";
-			if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-				$sql .= ", accountancy_code_buy = '" . $this->db->escape($this->accountancy_code_buy) . "'";
-				$sql .= ", accountancy_code_sell= '" . $this->db->escape($this->accountancy_code_sell) . "'";
+			if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+				if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+					$sql .= ", accountancy_code_buy = '" . $this->db->escape($this->accountancy_code_buy) . "'";
+					$sql .= ", accountancy_code_sell= '" . $this->db->escape($this->accountancy_code_sell) . "'";
+				}
 			}
 			$sql .= ",webservices_url = ".(!empty($this->webservices_url) ? "'".$this->db->escape($this->webservices_url)."'" : "null");
 			$sql .= ",webservices_key = ".(!empty($this->webservices_key) ? "'".$this->db->escape($this->webservices_key)."'" : "null");
@@ -1533,24 +1541,26 @@ class Societe extends CommonObject
 				$action = 'update';
 
 				// update accountancy for this entity
-				if (!$error && !empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-					$this->db->query("DELETE FROM " . MAIN_DB_PREFIX . "societe_perentity WHERE fk_soc = " . $this->id . " AND entity = " . $conf->entity);
+				if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+					if (!$error && !empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+						$this->db->query("DELETE FROM " . MAIN_DB_PREFIX . "societe_perentity WHERE fk_soc = " . $this->id . " AND entity = " . $conf->entity);
 
-					$sql = "INSERT INTO " . MAIN_DB_PREFIX . "societe_perentity (";
-					$sql .= " fk_soc";
-					$sql .= ", entity";
-					$sql .= ", accountancy_code_buy";
-					$sql .= ", accountancy_code_sell";
-					$sql .= ") VALUES (";
-					$sql .= $this->id;
-					$sql .= ", " . $conf->entity;
-					$sql .= ", '" . $this->db->escape($this->accountancy_code_buy) . "'";
-					$sql .= ", '" . $this->db->escape($this->accountancy_code_sell) . "'";
-					$sql .= ")";
-					$result = $this->db->query($sql);
-					if (!$result) {
-						$error++;
-						$this->error = 'ErrorFailedToUpdateAccountancyForEntity';
+						$sql = "INSERT INTO " . MAIN_DB_PREFIX . "societe_perentity (";
+						$sql .= " fk_soc";
+						$sql .= ", entity";
+						$sql .= ", accountancy_code_buy";
+						$sql .= ", accountancy_code_sell";
+						$sql .= ") VALUES (";
+						$sql .= $this->id;
+						$sql .= ", " . $conf->entity;
+						$sql .= ", '" . $this->db->escape($this->accountancy_code_buy) . "'";
+						$sql .= ", '" . $this->db->escape($this->accountancy_code_sell) . "'";
+						$sql .= ")";
+						$result = $this->db->query($sql);
+						if (!$result) {
+							$error++;
+							$this->error = 'ErrorFailedToUpdateAccountancyForEntity';
+						}
 					}
 				}
 
@@ -1644,10 +1654,12 @@ class Societe extends CommonObject
 		$sql .= ', s.fk_effectif as effectif_id';
 		$sql .= ', s.fk_forme_juridique as forme_juridique_code';
 		$sql .= ', s.webservices_url, s.webservices_key, s.model_pdf';
-		if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-			$sql .= ', s.accountancy_code_buy, s.accountancy_code_sell';
-		} else {
-			$sql .= ', spe.accountancy_code_buy, spe.accountancy_code_sell';
+		if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+			if (empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+				$sql .= ', s.accountancy_code_buy, s.accountancy_code_sell';
+			} else {
+				$sql .= ', spe.accountancy_code_buy, spe.accountancy_code_sell';
+			}
 		}
 		$sql .= ', s.code_client, s.code_fournisseur, s.code_compta, s.code_compta_fournisseur, s.parent, s.barcode';
 		$sql .= ', s.fk_departement as state_id, s.fk_pays as country_id, s.fk_stcomm, s.mode_reglement, s.cond_reglement, s.transport_mode';
@@ -1672,8 +1684,10 @@ class Societe extends CommonObject
 			$sql .= ', sr.remise_client, sr2.remise_supplier';
 		}
 		$sql .= ' FROM '.MAIN_DB_PREFIX.'societe as s';
-		if (!empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
-			$sql .= " LEFT JOIN " . MAIN_DB_PREFIX . "societe_perentity as spe ON spe.fk_soc = s.rowid AND spe.entity = " . ((int) $conf->entity);
+		if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+			if (!empty($conf->global->MAIN_COMPANY_PERENTITY_SHARED)) {
+				$sql .= " LEFT JOIN " . MAIN_DB_PREFIX . "societe_perentity as spe ON spe.fk_soc = s.rowid AND spe.entity = " . ((int)$conf->entity);
+			}
 		}
 		$sql .= ' LEFT JOIN '.MAIN_DB_PREFIX.'c_effectif as e ON s.fk_effectif = e.id';
 		$sql .= ' LEFT JOIN '.MAIN_DB_PREFIX.'c_country as c ON s.fk_pays = c.rowid';
@@ -1837,7 +1851,7 @@ class Societe extends CommonObject
 				$this->client = $obj->client;
 				$this->fournisseur = $obj->fournisseur;
 
-				$this->note = $obj->note_private; // TODO Deprecated for backward comtability
+				$this->note = $obj->note_private; // TODO Deprecated for backward compatibility
 				$this->note_private = $obj->note_private;
 				$this->note_public = $obj->note_public;
 				$this->model_pdf = $obj->model_pdf;
@@ -1849,8 +1863,10 @@ class Societe extends CommonObject
 				$this->webservices_url = $obj->webservices_url;
 				$this->webservices_key = $obj->webservices_key;
 
-				$this->accountancy_code_buy     = $obj->accountancy_code_buy;
-				$this->accountancy_code_sell    = $obj->accountancy_code_sell;
+				if (!empty($conf->global->ACCOUNTANCY_USE_PRODUCT_ACCOUNT_ON_THIRDPARTY)) {
+					$this->accountancy_code_buy = $obj->accountancy_code_buy;
+					$this->accountancy_code_sell = $obj->accountancy_code_sell;
+				}
 
 				$this->outstanding_limit		= $obj->outstanding_limit;
 				$this->order_min_amount			= $obj->order_min_amount;
